pyproject.toml
[project]
name = "todomd"
version = "0.1.0"
description = ""
authors = [
    {name = "Leonardo Sá",email = "leo@leo-sa.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyairtable (>=3.0.2,<4.0.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "pyyaml (>=6.0,<7.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
